Demo - OpenAI/requirements.txt
flask
elasticsearch==8.18.1
openai
python-dotenv